requirements.txt
# Core
flask>=2.3.0
flask-cors>=4.0.0
flask-jwt-extended>=4.5.0
python-dotenv>=1.0.0

# Google Cloud
firebase-admin>=6.0.0
google-cloud-translate>=3.11.0
google-cloud-speech>=2.0.0
google-cloud-texttospeech>=2.14.0
google-api-python-client>=2.88.0
google-genai>=0.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
responses>=0.23.0
